refactor(security): split rules.rs 944 LOC en 5 sous-modules par catégorie
Le fichier monolithique rules.rs est éclaté en:
- rules/helpers.rs        (68 LOC)  helpers partagés (crate-private)
- rules/traffic.rs       (325 LOC)  volume, bounce, codes SMTP, queue, hourly
- rules/geo.rs           (196 LOC)  pays/ASN nouveaux ou high-risk
- rules/auth.rs          (117 LOC)  brute force + stale api key
- rules/identity.rs      (224 LOC)  SPF/DKIM drift, rotation, cross-tenant, rate-limit
- rules.rs (mod.rs)       (94 LOC)  re-exports publics + evaluate_all

Bénéfices clean code:
- Aucun fichier > 325 LOC (vs 944 avant)
- Regroupement par domaine métier (facilite l'ajout de règles)
- API publique inchangée: pub use préserve tous les call-sites
- Helpers privés isolés (pub(crate) au lieu de pollution globale)
